converted I_NODE_NAME and iNodeName and similar to byte and renamed

+++ lib/global.h
@@ -17,7 +17,7 @@ extern short relayPins[];  // an array of pin numbers to which relays are attach
 
 extern const bool MY_DEBUG;
 
-extern const short B_NODE_NAME;                                                   // BOX BY BOX
+extern const byte B_NODE_NAME;                                                   // BOX BY BOX
 extern const short I_DEFAULT_MASTER_NODE_NAME;                                    // BOX BY BOX
 extern const bool MESH_ROOT;                                                      // BOX BY BOX
